Undraft retreive patient test.
